requirements.txt
asgiref==3.5.0
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
drf-extra-fields==3.4.0
drf-writable-nested==0.6.3
Pillow==9.1.0
pytz==2022.1
sqlparse==0.4.2
